+++ packages/jest-dom-mocks/CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+### Added
+
+- New document.cookie mock
+
 ## [2.7.1] - 2019-07-03
 
 ### Fixed